processor/requirements.txt
# IFC Processor Service - Python Dependencies

# FastAPI web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# IFC processing
ifcopenshell==0.7.0

# Task queue
celery[redis]==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Security
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
